# web/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CompiledSection": {
        "properties": {
          "label": {
            "title": "Label",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "order": {
            "title": "Order",
            "type": "integer"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          }
        },
        "required": [
          "name",
          "label",
          "prompt",
          "roles",
          "order"
        ],
        "title": "CompiledSection",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReviewField": {
        "properties": {
          "current": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current"
          },
          "decided": {
            "title": "Decided",
            "type": "boolean"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/ReviewEntry"
            },
            "title": "Entries",
            "type": "array"
          },
          "enum_values": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enum Values"
          },
          "field_path": {
            "title": "Field Path",
            "type": "string"
          },
          "flagged": {
            "title": "Flagged",
            "type": "boolean"
          },
          "help_text": {
            "title": "Help Text",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "required": {
            "title": "Required",
            "type": "boolean"
          },
          "review_order": {
            "title": "Review Order",
            "type": "integer"
          },
          "section": {
            "title": "Section",
            "type": "string"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "value_type": {
            "title": "Value Type",
            "type": "string"
          }
        },
        "required": [
          "field_path",
          "label",
          "section",
          "value_type",
          "unit",
          "required",
          "help_text",
          "enum_values",
          "review_order",
          "entries",
          "current",
          "decided",
          "flagged"
        ],
        "title": "ReviewField",
        "type": "object"
      },
      "ReviewSection": {
        "properties": {
          "label": {
            "title": "Label",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "order": {
            "title": "Order",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "label",
          "order"
        ],
        "title": "ReviewSection",
        "type": "object"
      },
...
      "ReviewTokenCreate": {
        "properties": {
          "reviewer_name": {
            "title": "Reviewer Name",
            "type": "string"
          },
          "tender_id": {
            "title": "Tender Id",
            "type": "string"
          }
        },
        "required": [
          "tender_id",
          "reviewer_name"
        ],
        "title": "ReviewTokenCreate",
        "type": "object"
      },
      "ReviewTokenOut": {
        "properties": {
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "reviewer_name": {
            "title": "Reviewer Name",
            "type": "string"
          },
          "tender_id": {
            "title": "Tender Id",
            "type": "string"
          },
          "token": {
            "title": "Token",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url",
          "token",
          "tender_id",
          "reviewer_name",
          "expires_at"
        ],
        "title": "ReviewTokenOut",
        "type": "object"
      },
      "ReviewVersion": {
        "properties": {
          "documents": {
            "items": {
              "$ref": "#/components/schemas/ReviewDocument"
            },
            "title": "Documents",
            "type": "array"
          },
          "issued_on": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issued On"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "version_no": {
            "title": "Version No",
            "type": "integer"
          }
        },
        "required": [
          "version_no",
          "kind",
          "issued_on",
          "documents"
        ],
        "title": "ReviewVersion",
        "type": "object"
      },
...
      "SnapshotOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "reviewer": {
            "title": "Reviewer",
            "type": "string"
          },
          "snapshot": {
            "additionalProperties": true,
            "title": "Snapshot",
            "type": "object"
          },
          "tender_id": {
            "title": "Tender Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tender_id",
          "reviewer",
          "created_at",
          "snapshot"
        ],
        "title": "SnapshotOut",
        "type": "object"
      },
      "TenderCreate": {
        "properties": {
          "agency": {
            "title": "Agency",
            "type": "string"
          },
          "external_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Ref"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "agency",
          "title"
        ],
        "title": "TenderCreate",
        "type": "object"
      },
...
      "TenderField": {
        "description": "One compiled field of a tender type, with where it comes from.",
        "properties": {
          "domain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain"
          },
          "enum_values": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enum Values"
          },
          "help_text": {
            "title": "Help Text",
            "type": "string"
          },
          "item_keys": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Keys"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "namespace": {
            "title": "Namespace",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "required": {
            "title": "Required",
            "type": "boolean"
          },
          "review_order": {
            "title": "Review Order",
            "type": "integer"
          },
          "section": {
            "title": "Section",
            "type": "string"
          },
          "subdomain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subdomain"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "value_type": {
            "title": "Value Type",
            "type": "string"
          }
        },
        "required": [
          "path",
          "namespace",
          "domain",
          "subdomain",
          "section",
          "label",
          "value_type",
          "unit",
          "required",
          "help_text",
          "enum_values",
          "item_keys",
          "review_order"
        ],
        "title": "TenderField",
        "type": "object"
      },
      "TenderOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "title": "Created By",
            "type": "string"
          },
          "current_version_no": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Version No"
          },
          "external_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Ref"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "issuing_agency": {
            "title": "Issuing Agency",
            "type": "string"
          },
          "slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tender_type": {
            "title": "Tender Type",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tender_type",
          "issuing_agency",
          "external_ref",
          "title",
          "slug",
          "status",
          "current_version_no",
          "created_at",
          "created_by"
        ],
        "title": "TenderOut",
        "type": "object"
      },
      "TenderReview": {
        "properties": {
          "can_complete": {
            "title": "Can Complete",
            "type": "boolean"
          },
          "decided": {
            "title": "Decided",
            "type": "integer"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/ReviewField"
            },
            "title": "Fields",
            "type": "array"
          },
          "issuing_agency": {
            "title": "Issuing Agency",
            "type": "string"
          },
          "required_undecided": {
            "title": "Required Undecided",
            "type": "integer"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/ReviewSection"
            },
            "title": "Sections",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tender_id": {
            "title": "Tender Id",
            "type": "string"
          },
          "tender_type": {
            "title": "Tender Type",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/ReviewVersion"
            },
            "title": "Versions",
            "type": "array"
          }
        },
        "required": [
          "tender_id",
          "tender_type",
          "title",
          "issuing_agency",
          "status",
          "versions",
          "sections",
          "fields",
          "decided",
          "total",
          "required_undecided",
          "can_complete"
        ],
        "title": "TenderReview",
        "type": "object"
      },
...
      "TenderSchemaOut": {
        "properties": {
          "fields": {
            "items": {
              "$ref": "#/components/schemas/TenderField"
            },
            "title": "Fields",
            "type": "array"
          },
          "pack": {
            "title": "Pack",
            "type": "string"
          },
          "schema_name": {
            "title": "Schema Name",
            "type": "string"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "string"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/CompiledSection"
            },
            "title": "Sections",
            "type": "array"
          },
          "tender_type": {
            "title": "Tender Type",
            "type": "string"
          }
        },
        "required": [
          "tender_type",
          "pack",
          "schema_name",
          "schema_version",
          "sections",
          "fields"
        ],
        "title": "TenderSchemaOut",
        "type": "object"
      },
...
      "TenderView": {
        "properties": {
          "current_version_no": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Version No"
          },
          "decided": {
            "title": "Decided",
            "type": "integer"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/FieldView"
            },
            "title": "Fields",
            "type": "array"
          },
          "tender_id": {
            "title": "Tender Id",
            "type": "string"
          },
          "tender_type": {
            "title": "Tender Type",
            "type": "string"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "tender_id",
          "tender_type",
          "current_version_no",
          "decided",
          "total",
          "fields"
        ],
        "title": "TenderView",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Tender Intelligence Engine",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/files-auth": {
      "get": {
        "description": "Asked by the proxy before it serves a stored file (a PDF or a page image): the file\nmust belong to a document of the review token's tender.",
        "operationId": "authorise_file_api_v1_files_auth_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-forwarded-uri",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Forwarded-Uri"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Authorise File",
        "tags": [
          "review"
        ]
      }
    },
...
    "/api/v1/review-tokens": {
      "post": {
        "description": "A review link for one tender and one reviewer, valid for 30 days. An earlier link\nfor the tender stops working. Not reachable with a review token or from outside.",
        "operationId": "create_review_token_api_v1_review_tokens_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-reviewer",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Reviewer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewTokenCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewTokenOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Review Token",
        "tags": [
          "review"
        ]
      }
    },
    "/api/v1/schemas/tender/{tender_type}": {
      "get": {
        "description": "The compiled field list of a tender type, in review order.",
        "operationId": "get_schema_api_v1_schemas_tender__tender_type__get",
        "parameters": [
          {
            "in": "path",
            "name": "tender_type",
            "required": true,
            "schema": {
              "title": "Tender Type",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenderSchemaOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Schema",
        "tags": [
          "tenders"
        ]
      }
    },
    "/api/v1/tenders": {
      "get": {
        "operationId": "list_tenders_api_v1_tenders_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TenderOut"
                  },
                  "title": "Response List Tenders Api V1 Tenders Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Tenders",
        "tags": [
          "tenders"
        ]
      },
      "post": {
        "description": "Create a tender. Its documents are added as versions.",
        "operationId": "create_tender_api_v1_tenders_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-reviewer",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Reviewer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenderCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenderOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Tender",
        "tags": [
          "tenders"
        ]
      }
    },
    "/api/v1/tenders/{tender_id}": {
      "get": {
        "operationId": "get_tender_api_v1_tenders__tender_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "tender_id",
            "required": true,
            "schema": {
              "title": "Tender Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenderOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Tender",
        "tags": [
          "tenders"
        ]
      }
    },
    "/api/v1/tenders/{tender_id}/complete-review": {
      "post": {
        "description": "Complete the review of the token's tender: refused while a required field has no\ndecision. Marks the token completed and the tender reviewed, and stores the current\nview as the snapshot the gold set is made from.",
        "operationId": "complete_review_api_v1_tenders__tender_id__complete_review_post",
        "parameters": [
          {
            "in": "path",
            "name": "tender_id",
            "required": true,
            "schema": {
              "title": "Tender Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Complete Review",
        "tags": [
          "review"
        ]
      }
    },
...
    "/api/v1/tenders/{tender_id}/review": {
      "get": {
        "description": "The tender as the reviewer sees it: every field once, with what each version says\nabout it and the entry to decide (the latest version that states the field).",
        "operationId": "get_tender_review_api_v1_tenders__tender_id__review_get",
        "parameters": [
          {
            "in": "path",
            "name": "tender_id",
            "required": true,
            "schema": {
              "title": "Tender Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenderReview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Tender Review",
        "tags": [
          "review"
        ]
      }
    },
...
    "/api/v1/tenders/{tender_id}/snapshot": {
      "get": {
        "description": "The final values of the tender's latest completed review.",
        "operationId": "get_snapshot_api_v1_tenders__tender_id__snapshot_get",
        "parameters": [
          {
            "in": "path",
            "name": "tender_id",
            "required": true,
            "schema": {
              "title": "Tender Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Snapshot",
        "tags": [
          "review"
        ]
      }
    },
...
    "/api/v1/tenders/{tender_id}/view": {
      "get": {
        "description": "The current view: per field, the canonical value from the latest version that set\nit, with that version's number and kind. Canonical facts only, never candidates.",
        "operationId": "get_view_api_v1_tenders__tender_id__view_get",
        "parameters": [
          {
            "in": "path",
            "name": "tender_id",
            "required": true,
            "schema": {
              "title": "Tender Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TenderView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get View",
        "tags": [
          "tenders"
        ]
      }
    }
  }
}